added bonus to makefile + lstnew

+++ srcs/ft_memchr.c
@@ -6,11 +6,13 @@
 /*   By: alemarch <alemarch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/18 12:36:33 by alemarch          #+#    #+#             */
-/*   Updated: 2021/11/18 13:44:37 by alemarch         ###   ########.fr       */
+/*   Updated: 2021/11/19 10:37:01 by alemarch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memchr(void *s, int c, unsigned long n)
+#include"libft.h"
+
+void	*ft_memchr(void *s, int c, size_t n)
 {
 	unsigned long	i;
