+++ srcs/ft_memmove.c
@@ -6,11 +6,13 @@
 /*   By: alemarch <alemarch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/18 11:29:33 by alemarch          #+#    #+#             */
-/*   Updated: 2021/11/18 14:43:35 by alemarch         ###   ########.fr       */
+/*   Updated: 2021/11/19 10:35:44 by alemarch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dest, void *src, unsigned long n)
+#include"libft.h"
+
+void	*ft_memmove(void *dest, void *src, size_t n)
 {
 	unsigned long	i;
